chore: 지도 관련 컴포넌트 schedule 폴더에서 map 폴더로 이동
