--- frontend/src/features/schedule/MapContent.tsx
@@ -1,16 +1,0 @@
-import { css } from "@emotion/react";
-
-const MapContent = () => {
-  return (
-    <div id="map-content" css={MapContentContainer}>
-      MapContent
-    </div>
-  );
-};
-
-export default MapContent;
-
-const MapContentContainer = css`
-  position: relative;
-  height: 100%;
-`;
